iot-data-generator/requirements.txt
kafka-python==2.2.11
pydantic==2.11.5
pydantic-settings==2.9.1